requirements.txt
asgiref==3.5.2
crispy-tailwind==0.5.0
Django==4.1.3
django-browser-reload==1.6.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-tailwind==3.4.0
djangorestframework==3.14.0
ez-setup==0.9
fontawesomefree==6.2.0
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
# pysqlite3==0.4.7
# pysqlite3-binary==0.4.7.post8
